+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ==================
 
   * KUSTOM-33 Change version info in admin panel
-  * KUSTOM-4  Fix C/O address sync from Kustom to Magento
+  * KUSTOM-4 Fix C/O address sync from Kustom to Magento
 
 11.0.12 / 2025-06-03
 ==================